+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 - Add URLParse Option to auth package. [#179](https://github.com/xmidt-org/argus/pull/179)
+- Use latest version of httpaux. [#180](https://github.com/xmidt-org/argus/pull/180)
 
 ## [v0.3.16]
 - Allow auth package client code to pass the basculehttp.OnErrorHTTPResponse option. [#174](https://github.com/xmidt-org/argus/pull/174)